% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.R
\name{compute_distance}
\alias{compute_distance}
\title{calcul de distance selon un ordre de villes.}
\usage{
compute_distance(G,cities,start_city)
}
\arguments{
\item{G}{: la matrice des distances.}

\item{cities}{: vecteur des villes à visiter.}

\item{start_city}{: la ville de début du voyage.}
}
\value{
La distance totale parcourue en commençant par 'start_city' et visitant 'cities' et terminant en 'start_city'.
}
\description{
calcul de distance selon un ordre de villes.
}
\examples{
G = matrix(runif(4*4,min=10,max=50),nrow=4,ncol=4)
G = G \%*\% t(G) # rendre G symétrique.
diag(G) = 0 # annuler la diagonale de G.
## la symétrie de G n'est pas nécessaire.
start_city = 1 # la ville de Départ.
cities = c(2,3,4)  # un ordre de ville à visiter.
distance = compute_distance(G,cities,start_city) # calcul de distance du tour.
}
