% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_distance_Rcpp}
\alias{compute_distance_Rcpp}
\title{calcul de distance selon un ordre de villes en Rcpp.}
\usage{
compute_distance_Rcpp(G,cities,start_city)
}
\arguments{
\item{G}{: la matrice des distances.}

\item{cities}{: vecteur des villes à visiter.}

\item{start_city}{: la ville de début du voyage.}
}
\value{
la distance totale parcourue
}
\description{
calcul de distance selon un ordre de villes en Rcpp.
}
\examples{
G = matrix(runif(4*4,min=10,max=50),nrow=4,ncol=4)
G = G \%*\% t(G) # rendre G symétrique
diag(G) = 0 # annuler la diagonale de G
## la symétrie de G n'est pas nécessaire.
start_city = 0 # la ville de Départ.
cities = c(1,2,3) # un ordre de ville à visiter.
distance = compute_distance_Rcpp(G,cities,start_city)
}
