% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.R
\name{.construct_C_S_k}
\alias{.construct_C_S_k}
\title{(Fonction cachée) Construire le vecteur (la liste [C(S{k}, m) + G[m,k]] pour tout m dans 'Subset')}
\usage{
.construct_C_S_k(C, Subset, k, G)
}
\arguments{
\item{C}{: la matrice tel que C[S,k] le coût min du chemin à partir de 1 et se termine au 
#          sommet k, passant les sommets de l'ensemble S exactement une fois.}

\item{Subset}{: Un sous groupe du groupe complet des villes.}

\item{k}{: La ville pour laquelle on veut calculer C[S-k,m] pour tout m dans Subset.}

\item{G}{: la matrice des distances.}
}
\value{
(la liste [C(S{k}, m) + G[m,k]] pour tout m dans 'Subset')
}
\description{
(Fonction cachée) Construire le vecteur (la liste [C(S{k}, m) + G[m,k]] pour tout m dans 'Subset')
}
