% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.R
\name{.generate_x_y_G}
\alias{.generate_x_y_G}
\title{(Fonction cachée) Génère une matrice des distance avec  et les coordonnées des villes aléatoirement.}
\usage{
.generate_x_y_G(n)
}
\arguments{
\item{n}{: le nombre des villes.}
}
\value{
Une matrice des distances et les coordonnées des villes.
}
\description{
(Fonction cachée) Génère une matrice des distance avec  et les coordonnées des villes aléatoirement.
}
