% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.R
\name{.generateDistances}
\alias{.generateDistances}
\title{(Fonction cachée) Génère une matrice de distance aléatoirement.}
\usage{
.generateDistances(n, minDistance, maxDistance)
}
\arguments{
\item{n}{: le nombre des villes.}

\item{minDistance}{: le min des distance à ne pas dépasser.}

\item{maxDistance}{: le max des distance à ne pas dépasser.}
}
\value{
Une matrice des distances.
}
\description{
(Fonction cachée) Génère une matrice de distance aléatoirement.
}
