% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.R
\name{.get_subsets}
\alias{.get_subsets}
\title{(Fonction cachée) Génère toutes les sous partitions de taille p d'un groupe de taille n.}
\usage{
.get_subsets(set, p)
}
\arguments{
\item{set}{: le groupe initial.}

\item{p}{: la taille des partitions qu'on veut extraire de 'set'.}
}
\value{
L'ensemble des partitions de taille p du groupe 'set'.
}
\description{
(Fonction cachée) Génère toutes les sous partitions de taille p d'un groupe de taille n.
}
