% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.R
\name{.search_min_C_S_k}
\alias{.search_min_C_S_k}
\title{(Fonction cachée) recherche du min dans un vecteur et son index.}
\usage{
.search_min_C_S_k(C_S_k)
}
\arguments{
\item{C_S_k}{: Une liste construit par la fonction '.construct_C_S_k'.}
}
\value{
Le minimum du vecteur et l'index du minimum.
}
\description{
(Fonction cachée) recherche du min dans un vecteur et son index.
}
