% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{greedy_TSP_max_all}
\alias{greedy_TSP_max_all}
\title{greedy_TSP_max_all}
\usage{
greedy_TSP_max_all(data)
}
\arguments{
\item{data}{matrix of positions for cities in (x,y)}
}
\value{
a vector of indices ("best" order of the cities to visit)
}
\description{
greedy_TSP_max_all
}
