% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{held_karp_Rcpp}
\alias{held_karp_Rcpp}
\title{L'algorithme de Held_karp.}
\usage{
held_karp_Rcpp(G,n)
}
\arguments{
\item{G}{: la matrice des distances.}

\item{n}{: le nombre de villes.}
}
\value{
La séquence des villes qui donnent la moindre en distance et la distance optimale.
}
\description{
L'algorithme de Held_karp.
}
\examples{
G = matrix(runif(4*4 , min = 10, max = 50),nrow=4)
G = G \%*\% t(G) # rendre G symétrique (n'est pas nécessaire.).
diag(G) = 0 # annuler la diagonale de G.
n = 4 # nombre de villes.
results = held_karp_Rcpp(G,n)
results['path_opt'] # l'ordre optimale des villes.
results['dist_opt'] # la distance totale optimale.
}
