% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.R
\name{naive_method}
\alias{naive_method}
\title{Test de toutes les permutations possibles.}
\usage{
naive_method(G,cities,start_city)
}
\arguments{
\item{G}{: la matrice des distances.}

\item{cities}{: vecteur des villes à visiter.}

\item{start_city}{: la ville de début du voyage.}
}
\value{
La séquence des villes qui donnent la moindre en distance.
}
\description{
Test de toutes les permutations possibles.
}
\examples{
G = matrix(runif(4*4 , min = 10, max = 50),nrow=4)
G = G \%*\% t(G) # rendre G symétrique (n'est pas nécessaire.).
diag(G) = 0 # annuler la diagonale de G.
start_city = 1 # la ville de Départ.
cities = c(2,3,4)  # un ordre de ville à visiter.
results = naive_method(G,cities,start_city)
results['path_optimal'] # l'ordre des villes optimales.
results['dist_optimal'] # la distance obtenue avec l'ordre des villes optimales.
}
