% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{naive_method_Rcpp}
\alias{naive_method_Rcpp}
\title{Test de toutes les permutations possibles en Rcpp.}
\usage{
naive_method_Rcpp(G,cities,start_city)
}
\arguments{
\item{G}{: la matrice des distances.}

\item{cities}{: vecteur des villes à visiter.}

\item{start_city}{: la ville de début du voyage.}
}
\value{
la séquence des villes qui donnent la moindre en distance.
}
\description{
Test de toutes les permutations possibles en Rcpp.
}
\examples{
G = matrix(runif(4*4 , min = 10, max = 50),nrow=4)
G = G \%*\% t(G) # rendre G symétrique (n'est pas nécessaire.)
diag(G) = 0 # annuler la diagonale de G
## la symétrie de G n'est pas nécessaire.
start_city = 0 # la ville de Départ.
cities = c(1,2,3)  # un ordre de ville à visiter.
results = naive_method_Rcpp(G,cities,start_city)
results['path_opt'] # l'ordre des villes optimales
results['dist_opt'] # la distance obtenue avec l'ordre des villes optimales.
}
