% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{plot.TSP}
\alias{plot.TSP}
\title{city plot}
\usage{
\method{plot}{TSP}(x, ..., data, main = "", value = "")
}
\arguments{
\item{x}{the proposed best tour}

\item{...}{Other parameters}

\item{data}{matrix of positions for cities in (x,y)}

\item{main}{title for the plot}

\item{value}{to plot the length value of the proposed tour}
}
\value{
plot 2D of the cities
}
\description{
plotting the cities and the proposed best tour
}
