% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.R
\name{test_compare}
\alias{test_compare}
\title{Comparaison entre la méthode naïve et Held-Karp.
       Affiche les chemins optimaux obtenus et les distances associées.}
\usage{
test_compare(n,minDistance,maxDistance)
}
\arguments{
\item{n}{: le nombre des villes.}

\item{minDistance}{: le min des distance à ne pas dépasser.}

\item{maxDistance}{: le max des distance à ne pas dépasser.}
}
\value{
Affichage des résultats.
}
\description{
Comparaison entre la méthode naïve et Held-Karp.
       Affiche les chemins optimaux obtenus et les distances associées.
}
