% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.R
\name{test_compare_plot}
\alias{test_compare_plot}
\title{Comparaison entre la méthode naïve et Held-Karp.
#      Affiche les chemins optimaux obtenus et les distances associées.}
\usage{
test_compare_plot(n)
}
\arguments{
\item{n}{: le nombre des villes.}
}
\value{
Affichage des résultats et un plot des villes et 
#       retourne les données utilisée (matrice G et les coordonnées des villes).
}
\description{
Comparaison entre la méthode naïve et Held-Karp.
#      Affiche les chemins optimaux obtenus et les distances associées.
}
