% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.R
\name{time_compare}
\alias{time_compare}
\title{Comparaison entre la méthode naïve et Held-Karp.
#      Affiche un plot du temps pris par les algorithmes en fonction de nombre des villes.}
\usage{
time_compare(min_cities,max_cities)
}
\arguments{
\item{min_cities}{: le nombre minimale des villes à ne pas dépasser.}

\item{max_cities}{: le nombre maximale des villes à ne pas dépasser.}
}
\value{
Affichage des résultats et un plot du temps pris par les algorithmes en fonction de nombre des villes et
#       retourne les données utilisées (matrice G et les coordonnées des villes).
}
\description{
Comparaison entre la méthode naïve et Held-Karp.
#      Affiche un plot du temps pris par les algorithmes en fonction de nombre des villes.
}
