% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tour_superSimple}
\alias{tour_superSimple}
\title{tour_superSimple}
\usage{
tour_superSimple(n)
}
\arguments{
\item{n}{an integer}
}
\value{
nothing interesting... cities from 1 to n
}
\description{
tour_superSimple
}
