\name{USCA312}
\alias{USCA312}
\docType{data}
\title{TSP -- 312 cities in the US and Canada}
\description{
This data set contains the distances between 312 cities in the US and Canada
as an object of class \code{TSP}.
}
\usage{data("USCA312")}
\format{
  An object of class \code{TSP}.  
}
%\details{
%}
\source{
John Burkardt, CITIES -- City Distance Datasets, 
\url{http://www.csit.fsu.edu/~burkardt/datasets/cities/cities.html} 
}
%\references{
%}
%\examples{
%data("USCA312")
%## maybe str(USCA312) ; plot(USCA312) ...
%}
\keyword{datasets}
