% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{tour_length}
\alias{tour_length}
\title{Tour length value}
\usage{
tour_length(tour, data)
}
\arguments{
\item{tour}{the proposed best tour}

\item{data}{matrix of positions for cities in (x,y)}
}
\value{
the tour length
}
\description{
Computing the tour length value
}
