% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_TSP-package.R
\docType{package}
\name{TSP-package}
\alias{TSP-package}
\title{TSP: Traveling Salesperson Problem (TSP)}
\description{
Basic infrastructure and some algorithms for the traveling salesperson problem (also traveling salesman problem; TSP). The package provides some simple algorithms and an interface to the Concorde TSP solver and its implementation of the Chained-Lin-Kernighan heuristic. The code for \href{https://www.math.uwaterloo.ca/tsp/concorde/}{Concorde} itself is not included in the package and has to be obtained separately.
}
\section{Key functions}{

\itemize{
\item \code{\link[=solve_TSP]{solve_TSP()}}
}
}

\references{
Michael Hahsler and Kurt Hornik. TSP -- Infrastructure for the traveling salesperson problem. Journal of Statistical Software, 23(2):1--21, December 2007. \doi{10.18637/jss.v023.i02}
}
\author{
Michael Hahsler
}
