% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_TSP-package.R
\docType{package}
\name{TSP-package}
\alias{TSP-package}
\title{TSP: Traveling Salesperson Problem (TSP)}
\description{
Basic infrastructure and some algorithms for the traveling salesperson problem (also traveling salesman problem; TSP). The package provides some simple algorithms and an interface to the Concorde TSP solver and its implementation of the Chained-Lin-Kernighan heuristic. The code for Concorde itself is not included in the package and has to be obtained separately. Hahsler and Hornik (2007) \doi{10.18637/jss.v023.i02}.
}
\section{Key functions}{

\itemize{
\item \code{\link[=solve_TSP]{solve_TSP()}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mhahsler/TSP}
  \item Report bugs at \url{https://github.com/mhahsler/TSP/issues}
}

}
\author{
\strong{Maintainer}: Michael Hahsler \email{mhahsler@lyle.smu.edu} (\href{https://orcid.org/0000-0003-2716-1405}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Kurt Hornik (\href{https://orcid.org/0000-0003-4198-9911}{ORCID}) [copyright holder]
}

}
\keyword{internal}
