\name{USCA312}
\alias{USCA312}
\alias{USCA50}
\docType{data}
\title{USCA312/USCA50 -- 312/50 cities in the US and Canada}
\description{
The \code{USCA312} dataset contains the distances between 312 cities in the US
and Canada as an object of class \code{TSP}.
\code{USCA50} is a subset of \code{USCA312} containing only the 
first 50 cities.
}
\usage{
data("USCA312")
data("USCA50")
}
\format{
  An object of class \code{TSP}.  
}
%\details{
%}
\source{
John Burkardt, CITIES -- City Distance Datasets, 
\url{http://www.csit.fsu.edu/~burkardt/datasets/cities/cities.html} 
}
%\references{
%}
%\examples{
%data("USCA312")
%## maybe str(USCA312) ; plot(USCA312) ...
%}
\keyword{datasets}
